Implement drawing, Add first person camera
With the IDrawable interface being used by custom classes, it becomes
much easier to draw objects.

The clients object is not drawn, but instead used for a first person
camera using its position and rotation. At the moment, it is buggy, with
physics affecting rotation somewhat and no turn cap.