Update NPS, Implement basic shooting
NPS has been updated to fix the collision callback issue.

Firing with left click will create a rocket infront of the player
traveling in the direction they are facing. Upon colliding with
something, it is destroied and creates an explosion on the server.
Currently, an explosion will only apply a force to client objects within
the radius, not doing damage, and explosions are not predicted by
clients, so movement is somehwta janky.