Update NPS, Add kill zone
NPS update includes getter for typeID.

A kill zone static object has been added, which when PlayerObject
collides with, will cause them to respawn. To implement this,
onCollision has been defined seperatly for server and client, with the
server checking the typeID for a kill zone. A kill zone has been added
to the bottom of the level so if a player falls, they will collide with
it and respawn.